\name{TP53Region}
\alias{TP53Region}
\title{
  Range around TP53
}
\description{
  Returns a \code{GRanges} object consisting of the TP53 coordinates in
  hg19. All coordinates in these data are relative to that region.
}
\usage{
TP53Region()
}
\value{
  A \code{GRanges} of the extents of the TP53 gene in hg19.
}
\author{
  Michael Lawrence
}

\examples{
TP53Region()
}
